Limiting the number of wifi interfaces to 3
When modifying the structure of the data base I have forget to edit the API especially the store_rssi_values function so I got some troubles.
Finally I have resolved the issue.
I have also limited the number of wifi interfaces in order to get the nearest wifi source and to reduce the mismatch of the length between the rssi strings stored in the table and the length of the rssi_vector (to avoid problems with the Euclidian_distance function).