Adding a Grid emulating the area
Each Grid consists of reference recorded points labeled in red circles that will be displayed automatically in the Grid once added using the record button (triggered by the scan function).
as well as a computed position associated with a different label that will be displayed once the where am I button is clicked.